requirements.txt
dill==0.3.1.1
gensim==3.8.3
h5py==2.10.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-bert==0.80.0
keras-embed-sim==0.7.0
keras-layer-normalization==0.14.0
keras-multi-head==0.22.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.0
keras-self-attention==0.41.0
keras-transformer==0.31.0
matplotlib==3.0.3
nltk==3.4.4
numpy==1.16.4
pandas==0.25.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.9.1
simplejson==3.16.0
six==1.12.0
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-addons==0.6.0
tensorflow-datasets==2.0.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-hub==0.6.0
tensorflow-metadata==0.21.1
tensorflow-probability==0.8.0
tqdm==4.41.1